+++ assets/src/markdoc/components/Button.tsx
@@ -6,7 +6,7 @@ import * as icons from '@pluralsh/design-system/dist/icons'
 import styled from 'styled-components'
 
 import { isExternalUrl } from 'markdoc/utils/text'
-import { useMarkdocContext } from 'markdoc/DocsContext'
+import { useMarkdocContext } from 'markdoc/MarkdocContext'
 
 function Button({
   href,
