+++ assets/src/markdoc/components/Heading.tsx
@@ -1,5 +1,5 @@
 import { LinksIcon } from '@pluralsh/design-system'
-import { useMarkdocContext } from 'markdoc/DocsContext'
+import { useMarkdocContext } from 'markdoc/MarkdocContext'
 import { Link } from 'react-router-dom'
 
 import styled, { useTheme } from 'styled-components'
