+++ assets/src/markdoc/components/Link.tsx
@@ -3,7 +3,7 @@ import type { ReactNode } from 'react'
 import styled from 'styled-components'
 
 import { isExternalUrl } from 'markdoc/utils/text'
-import { useMarkdocContext } from 'markdoc/DocsContext'
+import { useMarkdocContext } from 'markdoc/MarkdocContext'
 
 function Link({
   href,
